// public/shared/interfaces.ts
import React from 'react'

export interface carrito {
  id: number | null;
  titulo: string;
  cantidad: number | null;
  imagen1: string;
  categoria: string;
  precio: string;
  url: string
}
export interface valuesTransaccion {
    id: number
    id_transaccion: number
    nombres: string
    apellidos: string
    status: string
    tipo: string
    order_id: string
    email: string
    celular: string
    comentario: string | null
    delivery: string
    total_pago: string
    array_productos: string
    estado: number
  }

export interface configuracionValues {
    telefono: string
    celular1: string
    celular2: string
    correo1: string
    correo2: string
    horario1: string
    horario2: string
    direccion: string
    facebook: string
    instagram: string
    twiter: string
    linkedin: string
    youtube: string
    whatsapp: string
  }

export interface Values {
    nombres: string
    celular: string
    email: string
    base_proyecto: string
    nombre_empresa: string
    historia_empresa: string
    principales_servicios: string
    colores: string
    referencias: string
    transmitir: string
  }

export interface ImagenState {
    archivo: File | null
    archivoName: string
  }

export interface ImagePreviewProps {
    url: string
    setUrl: React.Dispatch<React.SetStateAction<string>>
    boton: boolean
    setBoton: React.Dispatch<React.SetStateAction<boolean>>
    setImagen: React.Dispatch<React.SetStateAction<ImagenState>>
    clase: string
  }

export interface ImagePreviewPropsUdpdate {
    globalUrl: string
    url: string
    setUrl: React.Dispatch<React.SetStateAction<string>>
    boton: boolean
    setBoton: React.Dispatch<React.SetStateAction<boolean>>
    imagen: string
    setImagen: React.Dispatch<React.SetStateAction<ImagenState>>
    clase: string
  }

export interface interfaceListaDiseño {
    id: number
    nombres: string
    celular: number
    email: string
    nombre_empresa: string
    created_at: string
    uptated_at: string
  }

// PAGINACION
export interface paginacionValues {
    totalPosts: number
    cantidadRegistros: number
    paginaActual: number
    setpaginaActual: (pagina: number) => void
  }

// DELETE
export interface deleteValues {
    ruta: string
    id: number
    token: string | null
    getData: () => Promise<void>
    totalPosts: number
    cantidadRegistros: number
    paginaActual: number
    setpaginaActual: (pagina: number) => void
  }

// BANNERS
export interface bannersValues {
    id: number
    imagen1: string
    created_at: string | null
    updated_at: string | null
  }

// MARCAS
export interface marcasValue {
    id: number
    imagen1: string
    imagen2: string
    created_at: string | null
    updated_at: string | null
  }

// CATEGORIAS
// LISTA
export interface categoriasValues {
    id: number
    nombre: string
    imagen1: string
    created_at: string | null
    updated_at: string | null
  }
  // CREACION - UPDATE
export interface categoriasValuesMoficate {
    nombre: string
    descripcion: string
  }

// PRODUCTOS
export interface productosValuesModificate {
    nombre: string
    idCategoria: string
    precio: string
    cantidad: string
    oferta: string
  }

// UPDATE-CREATE
export interface segundaSeccionValuesModificate {
    titulo: string
    descripcion: string
  }

export interface valoresValues {
    titulo: string
  }

export interface mapaValues {
    mapa: string
    mapa2: string
  }

export interface editorValues {
    content: string
    setContent: React.Dispatch<React.SetStateAction<string>>
  }

export interface arrayValues {
    id: number | null
    medida: string
    precio: string
    cantidad: string
    oferta: string
  }

// PRODUCTOS
export interface productosValues {
    id: number
    nombre: string
    categoria: string
    imagen1: string
    created_at: string | null
    updated_at: string | null
  }